% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_complaints_by_sector}
\alias{get_complaints_by_sector}
\title{Wrangle `Поплаки по СВР`}
\usage{
get_complaints_by_sector(path, lang)
}
\arguments{
\item{path}{path (local or url) to the excel file}

\item{lang}{A 2-letter language code, "mk", "sq", or "en"}
}
\description{
Wrangle `Поплаки по СВР`
}
\details{
This is the fourth sheet in the quarterly report in excel file
published by the Macedonian Ministry of Internal Affairs
}
\examples{
path <- "../data-raw/semejno-za-open-data-od-januari-do-mart-2022.xlsx"
get_complaints_by_sector(path = path, lang = "sq")
get_complaints_by_sector(path = path, lang = "mk")
get_complaints_by_sector(path = path, lang = "en")

}
